+++ lib/interact.dart
@@ -1,5 +1,7 @@
 library interact;
 
+import 'src/framework/framework.dart' show Context;
+
 export 'src/confirm.dart';
 export 'src/input.dart';
 export 'src/multi_progress.dart';
@@ -11,3 +13,6 @@ export 'src/select.dart';
 export 'src/sort.dart';
 export 'src/spinner.dart';
 export 'src/theme/theme.dart';
+
+/// Resets the Terminal to default values.
+void Function() reset = Context.reset;